# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CineConnect API
  version: 1.0.0
  description: |
    Source-first contract for CineConnect. Paths are added by vertical slice and
    must pass backend response-contract tests before they are marked complete.
servers:
  - url: https://cine.nalexustechnologies.com/api/v1
    description: Production
  - url: http://127.0.0.1:5000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /skills:
    get:
      tags: [Projects]
      operationId: listSkills
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Active skill catalog for project requirements and matching.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillListEnvelope"
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      responses:
        "200":
          description: Projects visible to the current user through active membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          description: Project created with the caller as owner/member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{project_id}:
    get:
      tags: [Projects]
      operationId: getProject
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail with members and requirements.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Projects]
      operationId: updateProject
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdateRequest"
      responses:
        "200":
          description: Project updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{project_id}/members:
    get:
      tags: [Projects]
      operationId: listProjectMembers
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active and inactive members for a visible project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMemberListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{project_id}/room:
    get:
      tags: [Projects]
      operationId: getProjectRoom
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project room aggregate with project, members, files, and room items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRoomEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{project_id}/room/items:
    post:
      tags: [Projects]
      operationId: createProjectRoomItem
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRoomItemCreateRequest"
      responses:
        "201":
          description: Project room item created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRoomItemEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{project_id}/files:
    get:
      tags: [Projects]
      operationId: listProjectFiles
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Files linked to a visible project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Projects]
      operationId: linkProjectFile
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileLinkRequest"
      responses:
        "201":
          description: Project file linked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileEnvelope"
        "200":
          description: Existing project file link updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFileEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{project_id}/requirements:
    get:
      tags: [Projects]
      operationId: listProjectRequirements
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Requirements for a visible project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRequirementListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Projects]
      operationId: createProjectRequirement
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequirementCreateRequest"
      responses:
        "201":
          description: Requirement created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRequirementEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /requirements/{requirement_id}:
    patch:
      tags: [Projects]
      operationId: updateProjectRequirement
      parameters:
        - in: path
          name: requirement_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRequirementUpdateRequest"
      responses:
        "200":
          description: Requirement updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRequirementEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Authentication is missing, expired, invalid, or revoked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: The caller is authenticated but lacks the required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationError:
      description: One or more request fields are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource was not found or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    SkillListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [skills]
          properties:
            skills:
              type: array
              items:
                $ref: "#/components/schemas/Skill"
        meta:
          $ref: "#/components/schemas/Meta"
...
    ProjectCreateRequest:
      type: object
      additionalProperties: false
      required: [title, project_type]
      properties:
        title:
          type: string
        project_type:
          type: string
        description:
          type: string
        city_id:
          type: [string, "null"]
        start_date:
          type: [string, "null"]
          format: date
        end_date:
          type: [string, "null"]
          format: date
        status:
          type: string
          enum: [draft, active, paused, completed, archived]
        estimated_budget_minor:
          type: [integer, "null"]
        currency:
          type: string
        visibility:
          type: string
          enum: [private, project_members]
        progress_percent:
          type: integer
    ProjectUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/ProjectCreateRequest"
    ProjectEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [project]
          properties:
            project:
              $ref: "#/components/schemas/Project"
        meta:
          $ref: "#/components/schemas/Meta"
    ProjectListEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [projects]
          properties:
            projects:
              type: array
              items:
                $ref: "#/components/schemas/Project"
        meta:
          $ref: "#/components/schemas/Meta"
...